Feature/ax 1704 parallel tests (#2227)
* not needed anymore

* CI blocks test

* parallel building on axonius instances with blocks

* better clean_dockers

* change ordering

* extra output

* some more output

* fix bug

* pylint fix i hope

* venv is already built in prepare_ci_env

* try executing jobs

* fix

* all ui + ut

* just a test - 1

* just a test - 2

* rookie mistake

* 3

* tests tests and more tests

* rookie mistake

* make everything more generic

* minor fixes

* some more potential bugfixes

* stupid bugfixes

* to be fixed later

* many improvements

* more

* Update test.py

* Update test.py

* stdout stderr jobs

* many tests

* regular no testsuite

* fix the bug

* more debug and bugfixes

* where are you, scheduling bug

* r

* remove aws get_all_regions, it takes too much time (~35 min) + add the tests printing

* make ut not fail everything as a temp hack.

* support 1 as rc from pytest

* ut with server

* Update test.py

* run ut tests chmod, clean dockers is better

* jobs have name + require set -e on scripts run + no debug messages

* junitxml for all

* run_pytest.py

* ut_tests.sh chmod

* transform to docker-py and add docker logs + logs artifacts

* Update test.py

* fix locations

* fix bug + extract xml from tar + report immediately to teamcity

* try without extracting

* set_image_attachment

* tired or what

* how

* generic location

* shortcut not working, revert to extracting from files

* allow printing output and getting artifacts even on error

* fix bug + flowid test

* streams support

* fix flows bugs

* make everything flowid and streamable

* fix bug + more debugging messages

* different directory for screenshots

* just a test

* fix paths

* do not export, see if it works

* make printing thread safe

* screenshot

* instance memory management moved to instancemanager

* better no-cache

* make another debugging optional param

* print the params

* fix thread-safe stdout for teamcity

* buffer

* untar xml and publish artifacts

* UI screenshots

* revoke test_code back to original

* formatting, except, crlf, requirements.txt checks are now in check_code

* build axonius-base-image

* sh original

* sys stderr in parallel_runner

* fix location for js logs as well

* max parallel for execute_docker_on_all

* no threaded print.

* the only threaded print

* flowid is str

* fix

* fix to comand

* fix tests

* xml dir relative

* show time

* bugs

* better messages, timeout

* more and more fixes

* why not running a discovery

* extract all screenshots

* the right overall time

* image tests

* try fixing things

* screenshots dir

* more logs

* print last seen if same device was yield twice or more

* print pngs

* fix test code

* make grid logs work

* raise outside of tc block

* fix the god damn git folder issue

* fix test_code

* test statistics summary

* test

* wild test

* recursive true + print the right thing

* fix bugs

* some more beauty

* make job descriptions better

* make paths generic

* fix bug

* remove dubug

* fix tests and show pylint mistakes

* disable test_pylint

* bug in alpine needs only one dns

* pylint

* remove broken reqs file

* fix iglob

* space

* try fixing the ui thing

* just a test

* only extract screenshot tar (do not save it) + publish screenshots + fix teamcity screenshot publishing

* refactoring to one artifacts folder

* fix bugs

* FIX BUG + TESTS

* fix two bugs

* support security groups

* just make exports be from new to old + fix bug

* bugfix

* update axonius host iamge

* make wave work already

* Always create network through python

* fix parallel

* fix printing exception

* remove bullshit

* make clean dockers fail on error

* some tests

* encryption key

* fix

* no string

* Update clean_dockers.sh

* flush stderr, to make that in one block

* create docker network in every system start, if it does not exist yet

* put some more debug messages

* decrease the time of timeout (becaue we have retry there)

* fix pylint and formatting

* try to stop gracefully

* nicer overall time

* no retry to avoid collection freeze

* some tries

* pylint

* exception handling

* some more debug

* jobs left print

* docker execute line is now with job name

* add timeout

* try fix

* better print and pylint

* prepare for cr

* cr

* cr

* make pics work

* unneeded

* fix have only custom network find

* try publishing afterwards because maybe we need to give tc time to parse

* shm size is 8 and not 4

* 16gb shm size

* memswap and tmpfs and wait 3 min

* Update test.py

* Update parallel_runner.py

* probably unneeded

* put a hell lot of

* remove screenshots

* try reducing container ram in gb size for mongodb purposes

* dynamic shm size

* soem more screenshots

* fix

* image attachment finally

* cr1

* cr2

* CR3

* cr4

* cr5

* mongodb gb ram refactoring

* increase timeout selenium

* cr6

* cr 7

* pylint and formatting

* sleep before putting image attachment

* test in times

* return back to 60

* try tuning performance some more

* try publishing the right ones

* remove tmpfs

* pylint

+++ devops/super_clean.sh
@@ -1,19 +1,25 @@
 #!/usr/bin/env bash
 
-echo "Removing all containers, volumes, networks, and axonius images"
+echo "Removing all containers"
 RUNNING_DOCKERS=$( docker ps -a -q )
 if [ "$RUNNING_DOCKERS" != "" ]; then
     docker rm -f ${RUNNING_DOCKERS}
+    docker rm -f ${RUNNING_DOCKERS}     # docker graph dependency
     RUNNING_DOCKERS=$( docker ps -a -q )
     if [ "$RUNNING_DOCKERS" != "" ]; then
+        echo "Running dockers remaining: $RUNNING_DOCKERS"
         exit 1
     fi
 fi
+
+echo "Removing all volumes"
 AVAILABLE_VOLUMES=$( docker volume ls -q )
 if [ "$AVAILABLE_VOLUMES" != "" ]; then
     docker volume rm -f ${AVAILABLE_VOLUMES}
+    docker volume rm -f ${AVAILABLE_VOLUMES}    # docker graph dependency
     AVAILABLE_VOLUMES=$( docker volume ls -q )
     if [ "$AVAILABLE_VOLUMES" != "" ]; then
+        echo "Available volumes remaining: $AVAILABLE_VOLUMES"
         exit 1
     fi
 fi
@@ -24,25 +30,38 @@ echo "Removing dangling docker images"
 DANGLING_IMAGES=$(docker images --filter "dangling=true" -q --no-trunc)
 if [ "$DANGLING_IMAGES" != "" ]; then
     docker rmi ${DANGLING_IMAGES}
+    docker rmi ${DANGLING_IMAGES}   # docker graph dependency
     DANGLING_IMAGES=$(docker images --filter "dangling=true" -q --no-trunc)
     if [ "$DANGLING_IMAGES" != "" ]; then
+        echo "Dangling images remaining: $DANGLING_IMAGES"
         exit 1
     fi
 fi
 
-AVAILABLE_IMAGES=$( docker images -q --filter=reference='axonius/*' )
-if [ "$AVAILABLE_IMAGES" != "" ]; then
-    docker rmi -f ${AVAILABLE_IMAGES}
+
+# Sometimes, we do not want the axonius images to get deleted.
+if [ "$1" != "--cache-images" ]; then
+    echo "Removing all images"
     AVAILABLE_IMAGES=$( docker images -q --filter=reference='axonius/*' )
     if [ "$AVAILABLE_IMAGES" != "" ]; then
-        exit 1
+        docker rmi -f ${AVAILABLE_IMAGES}
+        docker rmi -f ${AVAILABLE_IMAGES}   # docker graph dependency
+        AVAILABLE_IMAGES=$( docker images -q --filter=reference='axonius/*' )
+        if [ "$AVAILABLE_IMAGES" != "" ]; then
+            echo "Axonius images remaining: $AVAILABLE_IMAGES"
+            exit 1
+        fi
     fi
 fi
-AVAILABLE_NETWORKS=$( docker network ls -q --filter name=axonius )
+
+echo "Removing all networks"
+AVAILABLE_NETWORKS=$( docker network ls -q --filter type=custom)
 if [ "$AVAILABLE_NETWORKS" != "" ]; then
-    docker network rm axonius
-    AVAILABLE_NETWORKS=$( docker network ls -q --filter name=axonius )
+    docker network rm ${AVAILABLE_NETWORKS}
+    docker network rm ${AVAILABLE_NETWORKS}
+    AVAILABLE_NETWORKS=$( docker network ls -q --filter type=custom)
     if [ "$AVAILABLE_NETWORKS" != "" ]; then
+        echo "Axonius networks remaining: $AVAILABLE_NETWORKS"
         exit 1
     fi
 fi
